requirements.txt
fastapi>=0.115
uvicorn>=0.34
httpx>=0.28
PyJWT>=2.10
python-dotenv>=1.0
